+++ start_server.sh
@@ -1,3 +1,3 @@
-export FLASK_APP=hello.py
+export FLASK_APP=flask_app.py
 export FLASK_DEBUG=1
 python -m flask run
